feat(PositiveInteger): add PositiveInteger type
Add `PositiveInteger` type that is returned with the factory function
`positiveInteger`, which accepts a number as value and passes it to
an optional behavior function that enforces constraints.

The default constraints are:

- round value to nearest integer
- set value as absolute value
- default 0 to 1
- default NaN to 1

The default behavior function can be exchanged with a custom behavior
function that is passed as the second parameter to the factory function.

The `PositiveInteger` type is unioned with `number` to allow for
native arithmetic operations to be performed on the type.